pyproject.toml
[project]
name = "agentscope-runtime"
version = "1.0.0a2"
description = "A production-ready runtime framework for agent applications, providing secure sandboxed execution environments and scalable deployment solutions with multi-framework support."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.13",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "openai",
    "pydantic>=2.11.7",
    "requests>=2.32.4",
    "agentscope>=1.0.6",
    "docker>=7.1.0",
    "redis>=6.0.0",
    "oss2>=2.19.1",
    "pydantic-settings>=2.9.1",
    "python-dotenv>=1.0.1",
    "kubernetes>=33.1.0",
    "shortuuid>=1.0.13",
    "celery[redis]>=5.3.1",
    "a2a-sdk>=0.3.0",
    "wuying-agentbay-sdk>=0.5.0",
]

[tool.setuptools]
packages = { find = { where = ["src"] } }
include-package-data = true

[tool.setuptools.package-data]
"agentscope_runtime.engine.deployers.utils.service_utils" = ["*.j2"]

[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project.scripts]
runtime-sandbox-mcp = "agentscope_runtime.sandbox.mcp_server:main"
runtime-sandbox-server = "agentscope_runtime.sandbox.manager.server.app:main"
runtime-sandbox-builder = "agentscope_runtime.sandbox.build:main"
runtime-fc-deploy = "agentscope_runtime.engine.deployers.cli_fc_deploy:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.3.5",
    "pytest-asyncio>=0.23.0",
    "pre-commit>=4.2.0",
    "jupyter-book>=1.0.4.post1,<2.0.0",
    "furo>=2025.7.19",
    "pytest-cov>=6.2.1",
    "fakeredis>=2.31.0",
    "sphinx-autoapi>=3.6.0",
    "pytest-mock>=3.15.1",
]

langgraph = [
    "langgraph>=0.5.3",
]

agno = [
    "agno>=2.0.0",
]
autogen = [
    "autogen-agentchat>=0.7.4",
    "autogen-ext[openai]>=0.7.4",
]
agentrun = [
    "alibabacloud-agentrun20250910>=2.0.1",
    "alibabacloud_tea_openapi>=0.4.0"
]

aliyun_tablestore_ext=[
    "tablestore-for-agent-memory>=1.1.0",
    "dashscope>=1.24.4",
    "langchain-community>=0.3.27",
]

memory-ext = [
    "reme-ai==0.1.9 ; python_full_version >= '3.12'",
    "mem0ai>=0.1.117"
]


deployment = [
    "alibabacloud-agentrun20250910>=2.0.1",
    "alibabacloud-oss-v2",
    "alibabacloud-bailian20231229>=2.6.0",
    "build",
    "setuptools>=40.8.0",
    "wheel",
    "steel-sdk>=0.1.0",
    "alibabacloud-credentials",
    "jinja2",
    "psutil",
    "shortuuid>=1.0.13",
    "PyYAML",
    "oss2>=2.19.1"
]

sandbox = [
    "jinja2",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.23.0",
    "wuying-agentbay-sdk>=0.5.0",
    "alibabacloud_ecd20200930",
    "alibabacloud_eds_aic20230930",
    "alibabacloud_appstream_center20210218",
    "aiohttp",
    "e2b-desktop>=2.0.0",
    "psutil"
]
